Create two Errors files instead of one in Errors project
